Move top app bars to a new Navigation section in the ADS demo
The demo's Layouts tab was hosting the top app bars alongside cards. Give
them their own Navigation tab, positioned after Layouts.

https://app.asana.com/1/137249556945/project/1202857801505092/task/1215793353260352
