+++ android-design-system/design-system/src/main/res/values/donottranslate.xml
@@ -23,6 +23,7 @@
     <string name="tab_title_text_input" translatable="false">Text Input</string>
     <string name="tab_title_dialogs" translatable="false">Dialogs</string>
     <string name="tab_title_layouts" translatable="false">Layouts</string>
+    <string name="tab_title_component_navigation" translatable="false">Navigation</string>
     <string name="tab_title_templates" translatable="false">Templates</string>
     <string name="tab_title_component_interactive" translatable="false">Interactive</string>
     <string name="tab_title_component_messaging" translatable="false">Messaging</string>
